rename multi-step-wizard to ventilation wizard
