+++ frontend/src/modules/app/OverlayScreen.tsx
@@ -18,7 +18,7 @@ import {
 import { MessageType } from '../../store/proto/types';
 
 import ModalPopup from '../shared/ModalPopup';
-import MultiStepWizard from './multi-step-wizard/MultiStepWizard';
+import MultiStepWizard from '../ventilation-wizard/MultiStepWizard';
 import { getScreenLockPopup, setScreenLockPopup } from './Service';
 import { commitRequest } from '../../store/controller/actions';
 import { getBackendConnected } from '../../store/connection/selectors';
